Finalizando o Endpoint POST cadastracliente com validações de CPF e CNPJ

+++ src/main/java/com/projarctf/acmegames/application/usecases/cliente/ValidaClienteUseCase.java
@@ -9,7 +9,7 @@ public class ValidaClienteUseCase {
     @Autowired
     ClienteService clienteService;
 
-    public boolean validaCliente(long numero) {
+    public boolean validaCliente(int numero) {
         return clienteService.clienteExists(numero);
     }
 }
