+++ src/main/java/com/projarctf/acmegames/domain/repository/IClienteRepository.java
@@ -5,7 +5,7 @@ import com.projarctf.acmegames.domain.model.cliente.Cliente;
 
 public interface IClienteRepository {
     List<Cliente> getClientes();
-    Cliente getClienteByNumero(long numero);
+    Cliente getClienteByNumero(int numero);
     boolean cadastraCliente(Cliente cliente);
 }
