+++ src/main/java/com/projarctf/acmegames/infrastructure/persistence/entity/ClienteEntity.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class ClienteEntity {
     @Id
     @Column(name = "numero")
-    private Long numero;
+    private int numero;
 
     private String nome;
     private String endereco;
